:sparkles: Extract test methods
